+++ src/common/server.ts
@@ -33,8 +33,9 @@ export class SmartBearMcpServer extends McpServer {
   private samplingSupported = false;
   private elicitationSupported = false;
   private clients: Client[] = [];
+  private enabledToolsets?: string[];
 
-  constructor() {
+  constructor(enabledToolsets?: string) {
     super(
       {
         name: MCP_SERVER_NAME,
@@ -49,6 +50,11 @@ export class SmartBearMcpServer extends McpServer {
       },
     );
     this.cache = new CacheService();
+    if (enabledToolsets) {
+      this.enabledToolsets = enabledToolsets
+        .split(",")
+        .map((s) => s.trim().toLowerCase());
+    }
   }
 
   getCache(): CacheService {
@@ -85,6 +91,9 @@ export class SmartBearMcpServer extends McpServer {
     this.clients.push(client);
     await client.registerTools(
       (params, cb) => {
+        if (!this.isToolEnabled(client, params.toolset)) {
+          return null;
+        }
         const toolName = this.getCapabilityName(client, params.title);
         const toolTitle = this.getCapabilityTitle(client, params.title);
         if (toolName.length > 64) {
@@ -283,9 +292,41 @@ export class SmartBearMcpServer extends McpServer {
     return `${client.capabilityPrefix}_${title.replace(/\s+/g, "_").toLowerCase()}`;
   }
 
+  /**
+   * The tool is enabled if:
+   * - No enabled toolsets are defined on the server, or
+   * - The toolset is included in the enabled toolsets list, or
+   * - The toolset is in the client's default list and there is at least one toolset enabled for the client
+   * @param client
+   * @param toolset
+   * @returns whether to register the tool based on enabled toolsets configuration
+   */
+  private isToolEnabled(client: Client, toolset: string): boolean {
+    if (!this.enabledToolsets) {
+      return true;
+    }
+    const toolsetName =
+      `${client.configPrefix}:${toolset.replace(/[\s\-_]/g, "")}`.toLowerCase();
+    if (this.enabledToolsets.includes(toolsetName)) {
+      return true;
+    }
+    const someToolsEnabledForClient = this.enabledToolsets?.some(
+      (ts) =>
+        ts.split(":")[0].toLowerCase() === client.configPrefix.toLowerCase(),
+    );
+    if (
+      someToolsEnabledForClient &&
+      client.defaultToolsets?.includes(toolset)
+    ) {
+      return true;
+    }
+    return false;
+  }
+
   private getDescription(params: ToolParams): string {
     const {
       summary,
+      toolset,
       useCases,
       examples,
       inputSchema,
@@ -294,6 +335,10 @@ export class SmartBearMcpServer extends McpServer {
     } = params;
 
     let description = summary;
+
+    if (toolset) {
+      description += `\n\n**Toolset:** ${toolset}`;
+    }
 
     if (inputSchema && inputSchema instanceof ZodObject) {
       let parameters = Object.keys(inputSchema.shape)